Update FilestackClient and Upload tests
